float

+++ models.go
@@ -64,7 +64,7 @@ type Order struct {
 type OrderItem struct {
 	ProductID uint    `json:"product_id"`
 	Name      string  `json:"name"`
-	Count     int     `json:"count"`
+	Count     float32 `json:"count"`
 	Type      string  `json:"type"`
 	Subtotal  float64 `json:"subtotal"`
 }
@@ -141,8 +141,8 @@ type CreateOrderRequest struct {
 }
 
 type CreateOrderItem struct {
-	ProductID uint `json:"product_id"`
-	Count     int  `json:"count"`
+	ProductID uint    `json:"product_id"`
+	Count     float32 `json:"count"`
 }
 
 type UpdateOrderRequest struct {
@@ -158,9 +158,9 @@ type PrinterRequest struct {
 	Items    []PrinterItem `json:"items"`
 }
 type PrinterItem struct {
-	Product string `json:"product"`
-	Count   int    `json:"count"`
-	Type    string `json:"type"`
+	Product string  `json:"product"`
+	Count   float32 `json:"count"`
+	Type    string  `json:"type"`
 }
 
 // Response structs
